% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_permutation_distribution}
\alias{compute_permutation_distribution}
\title{Compute permutation distribution}
\usage{
compute_permutation_distribution(statistic, m, n, n_perm)
}
\description{
Used for precomputing permutation distribution of distribution free statistics
}
