% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate.unconditional}
\alias{evaluate.unconditional}
\title{Unconditional evaluation}
\usage{
evaluate.unconditional(
  m,
  n,
  limit,
  is_upper_limit,
  shifts,
  distribution_key,
  distribution_parameters,
  monitoring_statistic_key,
  monitoring_statistic_parameters,
  nsim,
  run_length_cap
)
}
\arguments{
\item{m}{The dimension used for the reference sample.}

\item{n}{The dimension used for the test samples.}

\item{distribution_key}{A string which identifies a distribution or a distribution family. Used in combination with the "distribution_parameters" parameter in order to select the IC distribution.}

\item{distribution_parameters}{A list with the distribution parameters (an empty list should be supplied if the distribution has not parameters).}

\item{monitoring_statistic_key}{A string used to select the monitoring statistic of the chart to be calibrated.}

\item{monitoring_statistic_parameters}{A list with the parameters to be supplied for the selected monitoring statistic.}

\item{nsim}{The number of simulations}

\item{run_length_cap}{A limit for the run length in the simulations used to guarantee convergence of the algorithm.}
}
\description{
Unconditional evaluation for the Stehwart-type charts implemented by this package
}
