% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_UCL}
\alias{find_UCL}
\title{Find UCL}
\usage{
find_UCL(reference_sample, n, target_ARL, nsim, nperm, test)
}
\arguments{
\item{x1}{An numeric vector}

\item{x2}{An numeric vector}

\item{B}{the number of permutations to be used for estimating the pvalue}
}
\description{
Compute approximated pvalue for the Tb test using a finite number of permutations.
}
