% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permtest.tab}
\alias{permtest.tab}
\title{Tab combined permutation test}
\usage{
permtest.tab(x1, x2, B)
}
\arguments{
\item{x1}{The first group's observations provided as a numeric vector.}

\item{x2}{The second group's observations provided as a numeric vector.}

\item{B}{The number of permutations used for estimating the pvalue}
}
\value{
The estimated permutation pvalue and the observed statistic
}
\description{
Estimate the right tail permutation pvalue for the Tab test using a finite number of permutations.
}
