% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permtest.tac}
\alias{permtest.tac}
\title{Tac permutation test}
\usage{
permtest.tac(x1, x2, B)
}
\arguments{
\item{x1}{An numeric vector}

\item{x2}{An numeric vector}

\item{B}{the number of permutations to be used for estimating the pvalue}
}
\description{
Compute pvalue estimate for the Tac test.
}
