% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permtest.tc}
\alias{permtest.tc}
\title{Tc permutation test}
\usage{
permtest.tc(x1, x2, B)
}
\arguments{
\item{x1}{An numeric vector}

\item{x2}{An numeric vector}

\item{B}{the number of permutations to be used for estimating the pvalue}
}
\description{
Compute approximated pvalue for the Tc test using a finite number of permutations.
}
