% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rl.conditional}
\alias{rl.conditional}
\title{Test UCL}
\usage{
rl.conditional(
  reference_sample,
  n,
  nsim,
  nperm,
  shifts,
  LCL,
  test,
  run_length_cap
)
}
\arguments{
\item{x1}{An numeric vector}

\item{x2}{An numeric vector}

\item{B}{the number of permutations to be used for estimating the pvalue}
}
\description{
Compute approximated pvalue for the Tb test using a finite number of permutations.
}
