% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbostat-pkg.R
\docType{package}
\name{turbostat-package}
\alias{turbostat}
\alias{turbostat-package}
\title{turbostat: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
