% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{turbostat.setseed}
\alias{turbostat.setseed}
\title{Ta permutation test}
\usage{
turbostat.setseed(s)
}
\arguments{
\item{x1}{An numeric vector}

\item{x2}{An numeric vector}

\item{B}{the number of permutations to be used for estimating the pvalue}
}
\description{
Compute approximated pvalue for the Ta test using a finite number of permutations.
}
