% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{turbostat.setseed}
\alias{turbostat.setseed}
\title{Set the seed used by this R packet}
\usage{
turbostat.setseed(seed)
}
\arguments{
\item{s}{The seed. Either an integer scalar or an integer vector of length 2 representing a 64-bit seed.}
}
\description{
Set the seed used by this R packet
}
